requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
nltk==3.8.1
textblob==0.17.1
google-cloud-storage==2.10.0
scipy==1.11.1
joblib==1.3.1
matplotlib==3.7.2
seaborn==0.12.2